Fixed issue where two-body operators were not Hermitian
Before when finding the basis of symmetry invariant two-body operators they would fall into three categories:
1.) a Hermitian two-body operator invariant to the symmetry operators provided
2.) Non-Hermitian two-body operator invariant to the symmetry operators
3.) Non-Hermitian two-body operator invariant to the symmetry operators and a transpose of the operators in 2.)
Since class 3.) operators were a transpose of one of the class 2.) operators, together the pair of 2.) and 3.) operators would form a symmetric   invariant and Hermitian operator.

Now all the two-body operators are symmetry invariant and Hermitian. They line up with the unique values in the random two-body Hamiltonian as well.